Cherry picked the PR from Beta Branch. (#33)
* Update README and bug fix (#24)

This commit updates the README file with information about AgeSQL installation, setup, and examples for a quick start. It also fixes a bug with the "select" variable, enabling successful program compilation.

* Added support for user defined function clauses (#26)

* Major bug fixes (#29)

* Update README and bug fix (#27)

This commit updates the README file with information about AgeSQL installation, setup, and examples for a quick start. It also fixes a bug with the "select" variable, enabling successful program compilation.

* Added support for user defined function clauses (#28)

* Update README and bug fix (#24)

This commit updates the README file with information about AgeSQL installation, setup, and examples for a quick start. It also fixes a bug with the "select" variable, enabling successful program compilation.

* Added support for user defined function clauses (#26)

---------

Co-authored-by: Carla Sanches <carlasanches8@gmail.com>

* Major bug fixes

---------

Co-authored-by: Carla Sanches <carlasanches8@gmail.com>

* Reformatted code; Updated README; Fixed bugs (#31)

---------

Co-authored-by: Carla Sanches <carlasanches8@gmail.com>
Co-authored-by: Ken Woon <xk.woon@gmail.com>